pyproject.toml
[project]
dependencies = [
  "google-cloud-aiplatform>=1.40,<2",
  "google-cloud-bigquery>=3.10,<4",
  "google-cloud-storage>=2.10,<3",
  "kfp>=2.5,<3",
  "protobuf>=4.25,<5",
  "fastapi>=0.110,<1",
  "uvicorn>=0.27,<1",
  "httpx>=0.27,<1",
  "pydantic>=2.5,<3",
  "pandas>=2.1,<3",
  "numpy>=1.26,<2",
  "pyyaml>=6.0,<7",
  "grpcio>=1.60,<2",
  "google-api-core>=2.15,<3",
  "googleapis-common-protos>=1.62,<2",
  "sqlalchemy>=2.0,<3",
  "pg8000>=1.30,<2",
  "cloud-sql-python-connector[pg8000]>=1.6,<2",
  "pydantic-settings>=2.1,<3",
  "dynaconf>=3.2,<4",
]
description = "I4G ML Platform — training, serving, evaluation, and monitoring for machine learning models."
name = "i4g-ml"
requires-python = ">=3.11"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.3",
  "black>=24.2",
  "isort>=5.13",
  "pre-commit>=3.6",
]
train-pytorch = [
  "torch>=2.1",
  "transformers>=4.38",
  "peft>=0.8",
  "datasets>=2.18",
  "evaluate>=0.4",
  "accelerate>=0.27",
]
train-xgboost = [
  "xgboost>=2.0",
  "scikit-learn>=1.4",
]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=68", "wheel"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
